JIRA SS-42 by RM-06
1. Added front end functionality for user management.
2. Common display messages added in displaymessage.properties.
3. New javascript file added to handle user management functions
(usermanagement.js).
4, New jsp files added usermanagement.jsp, userdetails.jsp,
userslist.jsp, useridoninvite.jsp.

+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/dao/UserDao.java
@@ -14,7 +14,7 @@ public interface UserDao extends GenericDao<User, Long> {
 
 	public List<User> fetchUsersBySimilarEmailId(User user, String emailId);
 	
-	public int getUsersCountForCompany(Company company);
+	public long getUsersCountForCompany(Company company);
 	
 	public List<User> getUsersForCompany(Company company);
